move dispatcher to registry
